hotshot_builder_core
Module builder_state

Structs

BuildBlockInfo

Response Message to be put on the response channel

BuilderState

BuiltFromProposedBlock

Builder State to hold the state of the builder

DAProposalInfo

DaProposalMessage

DA Proposal Message to be put on the da proposal channel

DecideMessage

Decide Message to be put on the decide channel

QCMessage

QC Message to be put on the quorum proposal channel

RequestMessage

Request Message to be put on the request channel

ResponseMessage

Response Message to be put on the response channel

Enums

MessageType

Unifies the possible messages that can be received by the builder

Status

Enum to hold the status out of the decide event

TransactionSource

Enum to hold the different sources of the transaction

TriggerStatus

Type Aliases

TxTimeStamp