hotshot_builder_core::builder_state
Function best_builder_states_to_extend

async fn best_builder_states_to_extend<Types: NodeType>(
    quorum_proposal: Arc<Proposal<Types, QuorumProposalWrapper<Types>>>,
    global_state: Arc<RwLock<GlobalState<Types>>>,
) -> HashSet<BuilderStateId<Types>>

best_builder_states_to_extend is a utility function that is used to in order to determine which BuilderStates are the best fit to extend from.

This function is designed to inspect the current state of the global state in order to determine which BuilderStates are the best fit to extend from. We only want to use information from GlobalState as otherwise we would have some insider knowledge unique to our specific BuilderState rather than knowledge that is available to all BuilderStates. In fact, in order to ensure this, this function lives outside of the BuilderState itself.

In an ideal circumstance the best BuilderState to extend from is going to be the one that is immediately preceding the [QuorumProposalWrapper] that we are attempting to extend from. However, if all we know is the view number of the [QuorumProposalWrapper] that we are attempting to extend from, then we may end up in a scenario where we have multiple BuilderStates that are all equally valid to extend from. When this happens, we have the potential for a data race.

The primary cause of this has to due with the interface of the ProxyGlobalState’s API. In general, we want to be able to retrieve a BuilderState via the [BuilderStateId]. The [BuilderStateId] only references a ViewNumber and a [VidCommitment] While this information is available in the [QuorumProposalWrapper], it only helps us to rule out BuilderStates that already exist. It does NOT help us to pick a BuilderState that is the best fit to extend from.

This is where the justify_qc comes in to consideration. The justify_qc contains the previous ViewNumber that is being extended from, and in addition it also contains the previous [Commitment<Leaf<Types>>] that is being built on top of. Since our BuilderStates store identifying information that contains this same leaf_commit we can compare these directly to ensure that we are extending from the correct BuilderState.

This function determines the best BuilderState in the following steps:

If we have a BuilderState that is already spawned for the current [QuorumProposalWrapper], then we should should return no states, as one already exists. This will prevent us from attempting to spawn duplicate BuilderStates.

Attempt to find all BuilderStates that are recorded within GlobalState that have matching view number and leaf commitments. There should only be one of these. But all would be valid extension points.

If we can’t find any BuilderStates that match the view number and leaf commitment, then we should return for the maximum stored view number that is smaller than the current [QuorumProposalWrapper].

If there is is only one BuilderState stored in the GlobalState, then we should return that BuilderState as the best fit.

If none of the other criteria match, we return an empty result as it is unclear what to do in this case.

Note: Any time this function returns more than a single entry in its HashSet result, there is a potential for a race condition. This is because there are multiple BuilderStates that are equally valid to extend from. This race could be avoided by just picking one of the entries in the resulting HashSet, but this is not done here in order to allow us to highlight the possibility of the race.