Trait hotshot_builder_core::builder_state::BuilderProgress

pub trait BuilderProgress<TYPES: NodeType> {
    // Required methods
    fn process_da_proposal<'life0, 'async_trait>(
        &'life0 mut self,
        da_msg: DaProposalMessage<TYPES>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn process_quorum_proposal<'life0, 'async_trait>(
        &'life0 mut self,
        qc_msg: QCMessage<TYPES>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn process_decide_event<'life0, 'async_trait>(
        &'life0 mut self,
        decide_msg: DecideMessage<TYPES>
    ) -> Pin<Box<dyn Future<Output = Option<Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn spawn_clone<'async_trait>(
        self,
        da_proposal: DAProposalInfo<TYPES>,
        quorum_proposal: Arc<Proposal<TYPES, QuorumProposal<TYPES>>>,
        leader: TYPES::SignatureKey,
        req_sender: BroadcastSender<MessageType<TYPES>>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn build_block<'life0, 'async_trait>(
        &'life0 mut self,
        matching_builder_commitment: VidCommitment,
        matching_view_number: TYPES::Time
    ) -> Pin<Box<dyn Future<Output = Option<BuildBlockInfo<TYPES>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn event_loop(self);
    fn process_block_request<'life0, 'async_trait>(
        &'life0 mut self,
        req: RequestMessage
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Trait to hold the helper functions for the builder

Required Methods

fn process_da_proposal<'life0, 'async_trait>( &'life0 mut self, da_msg: DaProposalMessage<TYPES> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

process the external transaction process the hotshot transaction process the DA proposal

fn process_quorum_proposal<'life0, 'async_trait>( &'life0 mut self, qc_msg: QCMessage<TYPES> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

process the quorum proposal

fn process_decide_event<'life0, 'async_trait>( &'life0 mut self, decide_msg: DecideMessage<TYPES> ) -> Pin<Box<dyn Future<Output = Option<Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

process the decide event

fn spawn_clone<'async_trait>( self, da_proposal: DAProposalInfo<TYPES>, quorum_proposal: Arc<Proposal<TYPES, QuorumProposal<TYPES>>>, leader: TYPES::SignatureKey, req_sender: BroadcastSender<MessageType<TYPES>> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait,

spawn a clone of builder

fn build_block<'life0, 'async_trait>( &'life0 mut self, matching_builder_commitment: VidCommitment, matching_view_number: TYPES::Time ) -> Pin<Box<dyn Future<Output = Option<BuildBlockInfo<TYPES>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

build a block

fn event_loop(self)

Event Loop

fn process_block_request<'life0, 'async_trait>( &'life0 mut self, req: RequestMessage ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

process the block request