hotshot_builder_core
Module service

Structs

BlockInfo

BlockSizeLimits

Adjustable limits for block size ceiled by maximum block size allowed by the protocol

GlobalState

GlobalState represents the internalized state of the Builder service as represented from its public facing API.

HandleReceivedTxns 🔒

HandleReceivedTxns is a struct that is used to handle the processing of the function handle_received_txns. In order to avoid the need to double allocate a Vec from processing these entries, this struct exists to be processed as an Iterator instead.

ProxyGlobalState

ReceivedTransaction

ReceivedTransaction represents receipt information concerning a received NodeType::Transaction.

Enums

AvailableBlocksError 🔒

AvailableBlocksError is an error enum that represents the class of possible errors that can be returned when calling available_blocks on a ProxyGlobalState. These errors are used for internal representations for consistency and testing, and do not leak beyond the ProxyGlobalState API. As such, they intentionally do not implement traits for serialization.

ClaimBlockError 🔒

ClaimBlockError is an error enum that represents the class of possible errors that can be returned when calling claim_block on a ProxyGlobalState. These errors are used for internal representations for consistency and testing, and do not leak beyond the ProxyGlobalState API. As such, they intentionally do not implement traits for serialization.

ClaimBlockHeaderInputError 🔒

GetChannelForMatchingBuilderError 🔒

GetChannelForMatchingBuilderError is an error enum that represents the class of possible errors that can be returned when calling get_channel_for_matching_builder_or_highest_view_builder on a GlobalState. These errors are used for internal representations for consistency and testing, and do not leak beyond the GlobalState API. As such, they intentionally do not implement traits for serialization.

HandleDaEventError 🔒

HandleDaEventError represents the internal class of errors that can occur when attempting to process an incoming da proposal event. More specifically these are the class of error that can be returned from handle_da_event_implementation.

HandleQuorumEventError 🔒

HandleQuorumEventError represents the internal class of errors that can occur when attempting to process an incoming quorum proposal event. More specifically these are the class of error that can be returned from handle_quorum_event_implementation.

HandleReceivedTxnsError 🔒

RecvError

An error returned from [Receiver::recv()].

TryRecvError

An error returned from [Receiver::try_recv()].

Constants

VID_RESPONSE_TARGET_MARGIN_DIVISOR 🔒

Functions

broadcast

Create a new broadcast channel.

handle_da_event 🔒

handle_da_event is a utility function that will attempt to broadcast the given da_proposal to the given da_channel_sender if the given details pass validation checks, and the [BroadcastSender] da_channel_sender is still open.

handle_da_event_implementation 🔒

handle_da_event_implementation is a utility function that will attempt to broadcast the given da_proposal to the given da_channel_sender if the given details pass all relevant checks.

handle_decide_event 🔒

handle_quorum_event 🔒

handle_quorum_event is a utility function that will attempt to broadcast the given quorum_proposal to the given quorum_channel_sender if the given details pass validation checks, and the [BroadcastSender] quorum_channel_sender is still open.

handle_quorum_event_implementation 🔒

Utility function that will attempt to broadcast the given quorum_proposal to the given quorum_channel_sender if the given details pass all relevant checks.

handle_received_txns 🔒

Utility function that will take the given list of transactions, txns, wraps them in a ReceivedTransaction struct and attempt to broadcast them to the given transaction [BroadcastSender] tx_sender. The broadcast itself it a non-blocking operation, and any failures of the broadcast are collected into the returned vector of Results.

run_non_permissioned_standalone_builder_service