hotshot_builder_core::service

Function handle_quorum_event

async fn handle_quorum_event<Types: NodeType>(
    quorum_channel_sender: &Sender<MessageType<Types>>,
    quorum_proposal: Arc<Proposal<Types, QuorumProposalWrapper<Types>>>,
    sender: <Types as NodeType>::SignatureKey,
)

handle_quorum_event is a utility function that will attempt to broadcast the given quorum_proposal to the given quorum_channel_sender if the given details pass validation checks, and the [BroadcastSender] quorum_channel_sender is still open.