hotshot_builder_core::service

Function handle_received_txns

pub(crate) async fn handle_received_txns<Types: NodeType>(
    tx_sender: &Sender<Arc<ReceivedTransaction<Types>>>,
    txns: Vec<Types::Transaction>,
    source: TransactionSource,
    max_txn_len: u64,
) -> Vec<Result<Commitment<<Types as NodeType>::Transaction>, BuildError>>

Utility function that will take the given list of transactions, txns, wraps them in a ReceivedTransaction struct and attempt to broadcast them to the given transaction [BroadcastSender] tx_sender. The broadcast itself it a non-blocking operation, and any failures of the broadcast are collected into the returned vector of Results.

There is also a max_txn_len parameter that is used to check to ensure that transactions that exceed this threshold will also not be broadcasted.