hotshot_builder_core::testing
Module basic_test

Re-exports

pub use crate::builder_state::BuilderState;

pub use crate::builder_state::MessageType;

pub use crate::builder_state::ResponseMessage;

Structs

BroadcastReceiver

The receiving side of a channel.

BroadcastSender

The sending side of the broadcast channel.

DaProposal

A proposal to start providing data availability for a block.

GeneralStaticCommittee

Dummy implementation of [Membership]

Leaf

This is the consensus-internal analogous concept to a block, and it contains the block proper, as well as the hash of its parent Leaf. NOTE: State is constrained to implementing BlockContents, is TypeMap::BlockPayload

Proposal

Prepare qc from the leader

QuorumProposal

Proposal to append a block.

SimpleCertificate

A certificate which can be created by aggregating many simple votes on the commitment.

SuccessThreshold

Defines a threshold which is 2f + 1 (Amount needed for Quorum)

ViewNumber

Type-safe wrapper around u64 so we know the thing we’re talking about is a view number.

Enums

RecvError

An error returned from Receiver::recv().

TryRecvError

An error returned from Receiver::try_recv().

Traits

BlockPayload

Abstraction over the full contents of a block

ConsensusTime

Trait for time compatibility needed for reward collection

NodeType

Trait with all the type definitions that are used in the current hotshot setup.

Functions

broadcast

Create a new broadcast channel.

Type Aliases

BLSPrivKey

BLS private key used to sign a message

BLSPubKey

BLS public key used to verify a signature

QuorumCertificate

Type alias for a QuorumCertificate, which is a SimpleCertificate of QuorumVotes