hotshot_builder_core
Module builder_state

Structs

BuildBlockInfo

Response Message to be put on the response channel

BuilderState

DAProposalInfo

DaProposalMessage

DA Proposal Message to be put on the da proposal channel

DecideMessage

Decide Message to be put on the decide channel

QuorumProposalMessage

Quorum proposal message to be put on the quorum proposal channel

RequestMessage

Request Message to be put on the request channel

ResponseMessage

Response Message to be put on the response channel

Enums

MessageType

Unifies the possible messages that can be received by the builder

Status

Enum to hold the status out of the decide event

TransactionSource

Enum to hold the different sources of the transaction

TriggerStatus

Constants

ALLOW_EMPTY_BLOCK_PERIOD 🔒

ALLOW_EMPTY_BLOCK_PERIOD is a constant that is used to determine the number of future views that we will allow building empty blocks for.

Functions

best_builder_states_to_extend 🔒

best_builder_states_to_extend is a utility function that is used to in order to determine which BuilderStates are the best fit to extend from.

Type Aliases

TxTimeStamp